BibKnowledge integration working
A-C-S-a invokes smart-replacement.  Try putting '08544' into an
affiliation box, hit the magic key sequence, and any right-minded
KB will give you 'Princeton U.'

There are a few things that keep this from being really finished, such as
the handcrafted test KB, or the fact that it's called "JoeTest".  But it's
working as a prototype, so I feel ok moving on to the next feature.